Retarget hardware config to RTX 4060 Laptop and document fork changes
- Switch dtype from float16 to bfloat16 across train/sample/bench/chat
  (Ada CC 8.9 supports bf16 natively, GradScaler no longer needed)
- Update estimate_mfu() denominator to 126 TFLOPS (RTX 4060 Laptop bf16)
- docs/study/00_overview.md: add missing doc 07 to the index, refresh the
  file tree with fork-only files, and add a section describing every
  difference from upstream karpathy/nanoGPT

+++ config/finetune_agent.py
@@ -1,6 +1,6 @@
 import time
 
-# AI Agent Expert Finetuning Config (RTX 2070 Optimized)
+# AI Agent Expert Finetuning Config (RTX 4060 Laptop Optimized)
 out_dir = 'out-agent-ft'
 eval_interval = 20
 eval_iters = 40
@@ -13,7 +13,7 @@ init_from = 'gpt2' # 124M model
 
 # Hardware Settings
 device = 'cuda'
-dtype = 'float16'
+dtype = 'bfloat16'
 compile = False
 
 # Training Settings
